% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradpsiT_sp.R
\name{gradpsiT_idea5}
\alias{gradpsiT_idea5}
\title{Gradient of the psiT function computed using sparse A. For this function,
\code{xAndBeta} should be a vector of length 2 * n, representing x_n and
beta_n}
\usage{
gradpsiT_idea5(xAndBeta, veccCondMeanVarObj, a, b, nu)
}
\description{
Gradient of the psiT function computed using sparse A. For this function,
\code{xAndBeta} should be a vector of length 2 * n, representing x_n and
beta_n
}
