% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvtrnd_cwrap.R
\name{mvtrnd_wrap}
\alias{mvtrnd_wrap}
\title{Wrapper of the \code{mvtrnd} function written in C++}
\usage{
mvtrnd_wrap(a, b, df, NN, veccObj, N, verbose = 0)
}
\description{
Wrapper of the \code{mvtrnd} function written in C++
}
