% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradpsiT_sp.R
\name{psiT_wrapper}
\alias{psiT_wrapper}
\title{psiT function that wraps the C function \code{psi}}
\usage{
psiT_wrapper(x, beta, nu, veccCondMeanVarObj, NN, a, b)
}
\description{
psiT function that wraps the C function \code{psi}
}
