% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{univar_order}
\alias{univar_order}
\title{Univariate variable reordering, described in Genz and Bretz (2009)
If failed due to PD singularity, the unfinished order will be returned 
and a warning will be issued}
\usage{
univar_order(a, b, sigma)
}
\arguments{
\item{a}{lower integration limits}

\item{b}{upper integration limits}

\item{sigma}{covariance matrix}
}
\value{
the new order
}
\description{
Univariate variable reordering, described in Genz and Bretz (2009)
If failed due to PD singularity, the unfinished order will be returned 
and a warning will be issued
}
