% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VeccTMVN.R
\name{VeccTMVN}
\alias{VeccTMVN}
\title{VeccTMVN}
\description{
Compute multivariate normal probabilities and sample from multivariate 
truncated normal distribution, taking advantage of the Vecchia approximation
}
\author{
jcao2416@gmail.com
}
